Use relaxed alarm restriction during rehoming
In this commit, the "--alarm-restrictions relaxed" option is included in
the sw-manager kube-rootca-update-strategy create to allow the
rehoming to proceed normally if the subcloud has no management
affecting alarms.

Test Plan:
PASS: Induce a non-management alarm condition on the subcloud prior to
rehoming the subcloud. Verify that the rehoming completes successfully.

Closes-Bug: 2020359

Change-Id: Iae2c085842a0243cab64cb7e4fbb56c0e0747659
Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>