Fix optimized restore ipv6 network issues
When running the optimized restore playbook with ipv6 network,
the ceph-mon process cannot be started because cannot bind to
the Controller-0 IP address. This is happening because when
restoring network configuration and bringing up the lo
interfaces, the ifup command is not overwriting the current
config if the interface is already up.

To solve this problem, ifup command has the '--force' parameter
to enforce the new configuration even with interfaces
already up.

Also, the parameter '--nowrap' was added to 'system addrpool-list'
commands to fix ipv6 optimized restore in case the address string
is too long.

Test-Plan:
  PASS: AIO-SX Optimized B&R IPv4
  PASS: AIO-SX Optimized B&R IPv6

Closes-bug: 2020584

Change-Id: I6003b810adf6ae655cb02938114dc2f79a4206fa
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>