Fix /var/lib/kubelet restore location
The extract directory argument for tar was in wrong location.
Running restore manually caused var/lib/kubelet to be extracted to wrong
place.  Running restore with dcmanager does not present an issue
because the tar command is run in a directory that allows it to work.

It appears that restoring var/lib/kubelet is not needed at all and
Kubernetes is able to be restored successfully without it.

TEST PLAN
PASS: Optimized restore AIO-SX
* With registry images
PASS: Optimized restore AIO-SX
* var/lib/kubelet deleted from platform backup
* With registry images
PASS: Optimized restore on AIO-SX subcloud using dcmanager
* With registry images

Closes-Bug: 2018613
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I22651814066ef24f7ad86b4dd02a20f1befc5309