Fix subcloud backup delete failure in 23.09
Currently subcloud backups taken in StarlingX 9.0
(23.09) cannot be removed as it's not in
supported_release_versions_for_backup list which
is 21.12, 22.06 and 22.12.

The change is to create a
minimum_supported_version_for_backup_delete which
will be used for validation instead.

PASS: Take a subcloud backup in 23.09 and verify
that the subcloud backup can be deleted.
PASS: Make a request to delete a subcloud backup
for an invalid release (e.g. 21.05), verify that
the playbook fails with the expected error
message.

Closes-Bug: 2023665

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: I7f44b8dd10331d3d8610eb4b5860e4e1bb54b044