Add error handling during task "create-account"
This commit is adding a better error handling message when
"create-account" task fails, and allows the user to re-run
the playbook with the same username, for situations of bad
password for example, it also deletes the ldap user.

Test Plan:

PASS: Run the playbook with invalid password and verify if
      proper message of error is shown.
PASS: After the test above, run the playbook again with same user
      and valid password and verify if the playbook finishes
      with no error.
PASS: Run the playbook and verify if the user is created.
PASS: SSH using the  newly created user.
PASS: Use 'mode=delete' and verify if keystone and ldap
      users were deleted after playbook execution.

Closes-Bug: 2024906
Change-Id: Id1d6a15a3248f2c26280d0c50a4a7e5a382d4438
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>