Use FQDN for MGMT network
The management network is used extensively for all internal
communication.
Since the original use of the network was a private network before it
was exposed for external communication in a distributed cloud
configuration, it was never designed to be reconfigured.

To support MGMT network reconfiguration the idea is to configure the
applications to use the hostname/FQDN instead of a static MGMT IP
address.

In this way, the MGMT network can be changed and the services and
applications will still work since they are using the hostname/FQDN
and the DNS will be responsible to translate to the current MGMT
IP address.
Given the complexities of supporting the multi-host reconfiguration,
the MGMT network reconfiguration will focus on support for AIO-SX only.

In a fresh install, the bringup initial services need to use hostname in
the barbican and fm config files in order to keep these services working
even when the management network is reconfigured.

Test plan ( Debian only )
 - AIO-SX and AIO-DX virtualbox installation IPv4/IPv6
 - Standard virtualbox installation IPv6
 - DC virtualbox installation IPv4 ( AIO-SX/DX subclouds )
 - AIO-SX and AIO-DX installation IPv4/IPv6
 - AIO-DX plus installation IPv6
 - DC IPv6 and subcloud AIO-SX
 - AIO-DX host-swact
 - DC IPv4 virtualbox with subcloud AIO-DX and AIO-DX
 - AIO-SX to AIO-DX migration
 - netstat -tupl ( no services are using the MGMT IP address )
 - Ran sanity/regression tests
 - Backup and Restore for AIO-SX/AIO-DX / DC subcloud AIO-SX
 - Standalone and DC upgrades will be tested in a separate task

Story: 2010722
Task: 48463
Depends-on: https://review.opendev.org/c/starlingx/config/+/886208

Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>

Change-Id: Idffbf5a90850f002287c4b583efab11fa30aaa34