Avoid unintended ldap configuration in subcloud
LDAP should not be configured in a subcloud as it connects to the
systemController's openldap service. The ldapserver_remote variable
is used to determine whether to configure a local ldap service or
connect to a remote one (to the system controller in this case).

Test plan:

PASS: Full build, install and bootstrap on a DC system and
      one subcloud added.
      Verify that the ldap directory is not configured in the subcloud,
      meaning that config.ldif is missing in /etc/ldap/schema.
      Create a user in the system controller and use it to login to
      the subcloud and verify that it works.
PASS: B&R in AIO-SX standalone and AIO-SX subcloud
PASS: Install standalone DX and SX environments and verify config.ldif
      is properly configured and ldap users work properly.
PASS: Run system host-reinstall in a DX environment.

Partial-Bug: 2025663

Change-Id: Iefc3dd65d0b68a6dcd86fe5c592dd98204f21086
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>