Fix k8s service-parameters not applied after restore
After restoring from a backup, the k8s service parameters
are not being applied to the k8s control plane components
and kubelet.

The service parameters for k8s are restored successfully
from the backup. However, the k8s custom configuration updating
script is not executed during this process. As a result, the
configuration of the Kubernetes control plane components and
kubelet (defaulting to bootstrap settings) does not align with
the restored k8s service parameters.

To resolve this, the updating script for k8s is now triggered
during the restore process. This ensures that the restored
configuration is applied to the k8s control plane components
and kubelet.

Changes on push_pull_local_registry.py file are necesary due
after applying the changes to the k8s control-plane (pods are
updated and restarted), the crictl repo is updated. This process
involves using Docker to initially pull images from the controller
repository and subsequently pushing them to the crictl repository.
Once the update is complete, an attempt is made to remove the
temporary images from the Docker repository. The modification
ensures that no failure occurs at this step while attempting to
remove an image that is still in use by any container.

Test Plan:
  PASS: Fresh Install AIO-SX.
  PASS: Add/modify and apply a custom K8s service-parameter.
  PASS: Create a backup.
  PASS: Restore system.
  PASS: Verify the service-parameters applied in K8s control plane.

  PASS: Fresh Install AIO-SX.
  PASS: Add/modify and apply OIDC service-parameters.
  PASS: Apply OIDC application.
  PASS: Create a backup.
  PASS: Restore system.
  PASS: Verify the service-parameters applied in K8s control plane.
  PASS: Verify the OIDC app status.

Closes-bug: 2025208

Change-Id: Ieda17f98552653fe8c7d2528c7a4f5ac9df65058
Signed-off-by: Jorge Saffe <jorge.saffe@windriver.com>