Fix image backup file generation for CentOS
Fix the issue that the subcloud backup create command doesn't create
user image registry backup file for the subcloud running CentOS when
the system controller is running Debian. The reason is the flag to
indicate if the user image file should be backed up was changed from
backup_user_local_registry (in 21.12) to backup_user_images
(in 22.12), so when the backup_user_images flag is passed from the
system controller to the subcloud, the subcloud backup playbook
doesn't recognize this new flag. The fix is mapping backup_user_images
to backup_user_local_registry internally in the backup playbook.

Test Plan:
Passed: Verify the user image registry backup file was successfully
        generated by running "dcmanager subcloud-backup create
        --local-only --registry-images" with a 21.12 subcloud on
        the 22.12 system controller.

Closes-Bug: 2027815

Signed-off-by: lzhu1 <li.zhu@windriver.com>
Change-Id: I38aff1d8048366251e2f35456c2647e8a2e25ceb