Fix restore failed by process admin-keystone on upgrade
During 21.12 -> 22.12 upgrade the playbook fails due the admin-keystone
is not found, this happens because the /etc/resolv.conf do not have
the right permission resulting on fail to bring up admin-keystone.
This commit check the permission of /etc/resolv.conf before the restart
of keystone service.

Test Plan:
AIO-SX upgrade 21.12 -> 22.12

Story: 2010651
Task: 48422

Change-Id: I0ef2198365cde82fcd02d42a56b8a47585ae06a6
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>