Fix: Exclude both snakeoil certs during restore
This commit is a follow-up update of commit:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/872208

There are actually two snakeoil certs.  Removing only one was not
enough.  To fix, the second snakeoil cert is now also excluded.
As part of the fix, the exclusions were moved to occur during
restore to allow more complete backups.

TEST PLAN
* Perform optimized B&R on AIO-SX
* Perform optimized B&R on AIO-SX with different images
  * Rebuilding the image without changes is OK for this

Closes-Bug: 2028407
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: Ibbf746e888f665c28466684d4d2e7a9135f3d6e4