Restore the previous backup after backup failure
Fix a typo that prevented the restore of previous
backup following a backup failure.

Test Plan:

1. PASS - Performed  a subcloud backup and created
a big file in central storage to simulate low storage.
Performed another subcloud backup and verified that
when the operation fails due to low disk space, the
previous backup is restored.

2. PASS - Performed a subcloud backup and verified
that temp directory is removed upon successful backup.

Closes-bug: 2028975

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: I47ae5899c764916d0daa6169e706aad49168d9cb