Change K8S pki folder perms and update /etc/hosts
To allow any user to create new K8S configuration files, the file
"/etc/kubernetes/pki/ca.crt" must be world readable and the cluster host
floating address must be present in "/etc/hosts".
The file "/etc/kubernetes/pki/ca.crt" already had the correct
permissions, but the parent folder didn't allow it to be read. The
folder permissions were changed from 700 to 755.
An entry with alias "controller-cluster-host", pointing to the cluster
host floating address, was added to "/etc/hosts".
Note: the upgrade scenario will be covered by another change and it is
tracked by task 48603 of storyboard 2010738.

Test Plan:

PASS: Execute the ansible bootstrap on an AIO-SX with this change
present. After the bootstrap is completed, verify that the permissions
of folder "/etc/kubernetes/pki" are set to "755" and that there is an
entry in "/etc/hosts" called "controller-cluster-host" pointing to the
cluster floating address.
PASS: Complete a DC deploy (duplex systemcontroller with 1 AIO-SX
subcloud) and verify on systemcontroller and subcloud controllers that
the permissions of folder "/etc/kubernetes/pki" are set to "755" and
that there is an entry in "/etc/hosts" called "controller-cluster-host"
pointing to the cluster floating address.

Story: 2010738
Task: 48564

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I6ec923ff0b5ce7c7bbc8443ab20ec63c0ae368cd