Remove CA install in SystemController (rehoming)
A redundancy inserted to fix the SystemController if the secret
'system-local-ca' was replaced but not installed as a trusted CA in
the SystemController was causing it to be overwhelmed in case of
parallel rehomings (the same certificate install call was being made
by several subclouds at the same time).

Since this install is not actually required (the CA is expected to
be installed as trusted CA), we are removing this step. Also added
a check to avoid running into a 250.001 alarm in the subcloud after
installing the certificate there.

Test Plan:
PASS: Rehome 10 subclouds in parallel.

Test Plan:
PASS: With two DX SystemControllers and 10 SX subclouds:
      1) Install and manage all the 10 subclouds in one of the
         SystemControllers;
      2) Unmanage the subclouds, check dates and if the subclous
         are free of alarms;
      3) In the other SystemController, migrate/rehome the 10
         subclouds in parallel.

Closes-Bug: 2029510

Change-Id: Ief56be4bdd0af71a8075aff21e6ebce6798959af
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>