Bootstrap Login Banner Configuration
The current banner customization requires the presence of
the custom banner files to be populated to /opt/banner
during the system bootstrap phase.

To remove this restriction, the bootstrap configuration
of the customer banner information has been changed to
permit the custom files to be specified as inline files
specified by the ansible overrides.

The contents of the files can be defined in the ansible
overrides as follows:
banner:
  issue:
    console: <multiline-string>
    network: <multiline-string>
  motd:
    head: <multiline-string>
    tail: <multiline-string>

Test Plan:
  PASS: Fresh Install SX env with custom banners
  PASS: Fresh Install STD env with custom banners
  PASS: Fresh Install DC central with custom banners
  PASS: Fresh Install DC subcloud with custom banners.

Story: 2010913
Task: 48762

Change-Id: I7389e62a8c96d3e219427f741e6d2557a6b04567
Signed-off-by: Jorge Saffe <jorge.saffe@windriver.com>