Recreate PVCs when restoring backup with wipe_ceph_osds=true
When restoring from a backup with the flag wipe_ceph_osds=true, the
recovered pods might get stuck waiting for the PVC to be attached and
mounted, but the ceph disks were wiped.

To recover from this situtation, a script is added to iterate through
each namespace checking for PVCs that were provisioned by ceph
provisioners. When a broken PVC is found, it is deleted and reapplied
to be reprovisioned. Then the Pods get back running.

Test Plan:
  PASS: AIO-SX Optimized B&R with helm charts using pvcs applied
  PASS: AIO-SX Legacy B&R with helm charts using pvcs applied
  PASS: Standard with Storage nodes B&R with helm charts using pvcs
        applied
  PASS: AIO-SX subcloud Optimized B&R with wipe_ceph_osds=true

Closes-bug: 2030120

Change-Id: I2a0a15c526689ec30eea514933a6d1eabebf23db
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>