Manually recover sysinv services before manifest when rehoming
When updating the sysinv user's password, there's a time window that
the sysinv conductor may fail to get openstack users due to
authentication failure as its password was changing.

This commit switches the apporach: firstly manually update and recover
the sysinv on the active controller. Once the sysinv on the active
controller is updated with the new sysinv password, it is to safe to
retry updating the sysinv password, which ensures the puppet manifest is
triggered to update the configuration on controller-1 if the system is
a duplex or multi-node.

Test Plan:
1 Rehome 5 AIOSX subclouds in a batch successfully
2 Rehome 5 AIODX subclouds in a batch successfully

Closes-bug: 2031182
Change-Id: Ib7bb90dbcf1d0e265df82ed39ac90a08793f126f
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>