Split upgrade data to improve upgrade speed
When upgrading from older releases that do not have the improved backup
playbook the upgrade data tar will be very large.  This large backup
creates a significant cost in terms of time as there are many tar
commands that act on it during the upgrade playbook.

So reduce the impact of this, near the beginning of the upgrade playbook
all required files are extracted from it and put into two new tars that
are used during the upgrade.

The first is a miniature tar, which contains only minimum required files
from /etc, /opt, and /var, excluding patches.

The second is one that includes only patch metadata.  The RPMs are
not included because the are not used on Debian.

The original upgrade data tar is left untouched allow revert.

TEST PLAN
PASS: stx6 to stx8 AIO-SX upgrade
PASS: stx6 to stx8 AIO-SX subcloud upgrade

Task: 2010798
Story: 48665
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I92fd20127614ae0788bcf9b6d9ea3a81a55e739a