Fix AIO-SX to AIO-DX with admin network
This commit fixes an issue seen when attempting to migrate from an
AIO-SX to AIO-DX subcloud using the admin network.

The current code assumes the network name of the admin network is
'admin'.  This is not necessarily the case, as the admin network
can be user created and can be named anything.

The solution is to determine the name of the network of type
'admin', and use that to determine the interface/network
association.

Testing:

- Perform AIO-SX to AIO-DX migration for a system with a user
  defined network name of type 'admin'

Story: 2010319
Task: 47563

Change-Id: Idad594494c8dbb0a8d8471a5ee2574cc2083a5d3
Signed-off-by: Steven Webster <steven.webster@windriver.com>