Wait alarm to clear in trusted CA cert install
Install of trusted CA certificate raises a 250.001 alarm
(config-out-date). The playbook role install_trusted_ca, in a normal
use case, should wait for the alarm to clear up before moving forward.

Added code to monitor the change of the .config_applied file, written
by puppet after executing the manifest to update current config, and
check if the alarm is cleared. This way, we know that the next config
is applied before checking if the alarm is cleared, avoiding the case
where the alarm is raised after we check and is still present when we
try to create or apply the Kubernetes RootCA Update strategy.

Included ignore_alarms flag to keep the current cert-manager migration
behavior.

Removed non-essential openldap and kubeapi service restarts that were
affecting rehoming. These restarts were workarounds for rare issues
that haven't yet been found in non-virtual environments, and can be
dealt in a proper way if reproduced in the future.

Test Plan:
PASS: Execute cert-manager migration playbook and observe if new
      trusted certs are correctly installed.
      Re-execute the migration with the same inventory and observe
      that it works even if the certs are already installed.

PASS: With two DX SystemControllers and 10 SX subclouds:
      1) Install and manage all the 10 subclouds in one of the
         SystemControllers;
      2) Unmanage the subclouds, check dates and if the subclouds
         are free of alarms;
      3) In the other SystemController, migrate/rehome the 10
         subclouds in parallel.

Closes-Bug: 2034346

Change-Id: I3ad2c4d7afa713a41db0cf431333e179f6d925ff
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>