Subcloud backup failure removes custom backup directory
This commit fixes the issue of deleting the custom
backup directory defined in the overrides file and
stored on a subcloud when the backup creation
process fails.

Test Plan:

PASS - Deploy a subcloud and create an override backup
file with a custom backup directory defined. Take a
subcloud backup passing the --local-only parameter
to store the backup in the subcloud. Create a custom
file with the same size of the remaining available
space in the custom backup directory. Take another
backup. Verify that the previous backup was not
removed when the backup creation failed due to a
lack of space.

PASS - Deploy a subcloud and create an override backup
file with a custom backup directory defined. Take a
subcloud backup and create a custom file with the same
size of the remaining available space in the custom
backup directory. Take another backup and verify
that the previous backup was not removed when the
backup creation failed due to a lack of space.

Closes-bug: 2033529

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: I82c3f6cb82d1947aa617e4202b5449abfc0bf735