Fix: Do not bring up Host floating IP (VIP) twice
On certain system configurations, cluster VIP will already be brought
up.  If the playbook attempts to bring it up a
second time there will be an error.

This can potentially occur during optimized restore or
Debian to Debian AIO-SX upgrades.  When the cluster-host-subnet's
floating IP is configured to be on the lo interface as
shown in `system addrpool-list`.

The second attempt now checks to ensure the address is not up already.

TEST PLAN
PASS: Optimized B&R on AIO-SX
* Run on affected system,
* Run on unaffected system
PASS: Upgrade stx6 to stx8 on AIO-SX
PASS: Optimized B&R on AIO-SX after stx6 to stx8 upgrade

Closes-Bug: 2034616
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I60cdcbc4fcf20dddf11a9af7dba8b4031ed006f8