Ensure affected app is removed over restore if ceph osds are wiped
When the controller unlocks, there is a chance platform-integ-apps has
not yet been completely removed. This causes the removal to break and
sets the state from removing to remove-failed. This removal is necessary
to ensure that the platform-integ-apps will be reapplied after the
unlock.

To ensure that the platform-integ-apps was removed over restore, a
new task was added to classic and optimized restore. It waits for the
platform-integ-apps to be removed successfully before finishing it.
This new task will be executed only if wipe_ceph_osds is true.

TEST PLAN:
PASS: Set wipe_ceph_osds=true and restore_mode=optimized in host
    overrides and verify that AIO-SX restore completes successfully with
    platform-integ-apps removed before the unlock and reapplied post
    unlock.
PASS: Set wipe_ceph_osds=false and restore_mode=optimized in host
    overrides and verify that AIO-SX restore completes successfully with
    platform-integ-apps in applied state before and after host unlock.

PASS: Set wipe_ceph_osds=true in host overrides and verify that AIO-SX
    restore completes successfully with platform-integ-apps removed
    before the unlock and reapplied post unlock.
PASS: Set wipe_ceph_osds=false in host overrides and verify that AIO-SX
    restore completes successfully with platform-integ-apps in applied
    state before and after host unlock.

PASS: Set wipe_ceph_osds=true in host overrides and verify that AIO-DX
    restore completes successfully with platform-integ-apps removed
    before the unlock and reapplied post unlock.
PASS: Set wipe_ceph_osds=false in host overrides and verify that AIO-DX
    restore completes successfully with platform-integ-apps in applied
    state before and after host unlock.

Closes-Bug: #2035115
Change-Id: I1f0b2d55ab1ddc7c4147cf815f343b99f7a2ef63
Signed-off-by: Guilherme Costa <guilherme.costa@windriver.com>