Bootstrap subclouds with TLS 'system-local-ca'
Included code to bootstrap subclouds with TLS type 'system-local-ca'
secret and create a local ClusterIssuer using this secret. Before,
subclouds were bootstrapped with 'system-local-ca' as an Opaque type
secret, and the type of this secret would change only if the user
performed cert-manager migration (secret would be TLS after migration).

This avoids deployments failing when users try to create their issuers
and certs using this secret in subclouds.

This code also handles upgrades for SX subclouds (DX uses an API call,
this API will be fixed in a separated review for 'config' repo).

Test Plan:
PASS: Check for TLS 'system-local-ca' in active controllers in:
        - Bootstraped AIO-SX and DC (SystemController + SX subcloud)
        - Upgraded AIO-SX and DC (SystemController + SX subcloud)
        - Rehomed SX subcloud (2 DC SystemController + SX subcloud)

Partial-Bug: 2029508

Change-Id: I47866b6cdf2f4abdf8b4d9f952f66d9b57ac6198
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>