Subcloud restore failed due to no images backup file
The default backup filename for user images does not match between
the backup and restore playbooks on the inactive load of the subcloud.
In version 21.12, the filename has the prefix
'docker_local_registry_backup', while in version 22.12, it uses
the prefix 'user_images_backup.' To resolve this, we need to replace
the old filename prefix with the new one and apply it to the 21.12
backup playbook. This change will ensure that on the 21.12 subcloud,
the backup filename aligns with the new prefix. Subsequently, when we
initiate the restore process from the 22.12 system controller,
the filename validation will pass, allowing for a successful restore.

Test Plan:
PASS: Successfully backed up and restored subcloud with inactive load
      by running the following commands:
      $ dcmanager subcloud-backup create --subcloud <subcloud_name>
      --sysadmin-password XXX --local-only --registry-images
      $ dcmanager subcloud-backup restore --subcloud <subcloud_name>
      --sysadmin-password Li69nux* --local-only --registry-images
      --with-install --release 21.12
PASS: Successfully backed up and restored subcloud with active load
PASS: Verified the image backup was used during restore

Closes-Bug: 2035125

Change-Id: I0c7cdd4033cdea57f3a4c95505f30c01c0dc50cd
Signed-off-by: lzhu1 <li.zhu@windriver.com>