Revert "Add support to pull intel-qat-plugin and intel-gpu-plugin images"
Reason for revert: The functionality added here will
be a part of a different application, and the code
here is now obsolete.

This reverts commit 880962:
Add support to pull intel-qat-plugin and intel-gpu-plugin images
https://review.opendev.org/c/starlingx/ansible-playbooks/+/880962

Reference to the application where these functionalities will be added: https://opendev.org/starlingx/app-intel-device-plugins
The same storyboard can be tracked for further updates.

Story: 2010604
Task: 48738

Change-Id: I7357f101bc75498f8681c3a7ae9d32144bafdf74
Signed-off-by: Aman Pandae <amanpandae.mothukuri@windriver.com>