When executing upgrade bootstrap for AIO-SX only update /etc/network/
During upgrade bootstrap, the runtime execution of the puppet class
platform::network::runtime created a lock timeout with sysinv-agent
to allow the interface configuration in the kernel. The lock exists
to allow sysinv-agent to collect interface information for the
system inventory.

The optimized upgrade feature created this runtime execution to fill
the contents in /etc/network/interface.d/ and /etc/network/routes
to be available during the network bringup phase that happens after
the system unlock in an earlier step than the regular puppet manifest
execution. But as part of apply_network_config.sh execution it also
brings up the interfaces, accessing the script protected section and
causing lock timeout.

This change adds the file /var/run/.network_upgrade_bootstrap to
indicate a upgrade bootstrap  is under way to just populate the
/etc/network files and to not activate the interfaces.

The interface activation is not needed as the bootstrap is still under
way and the minimal network configuration is already provided prior to
the bootstrap. After unlocking the files in /etc/network will provide
a faster network availability as systemd's network service is one of
the first to be executed during boot.

The new flag will be used in
https://review.opendev.org/c/starlingx/stx-puppet/+/895726

Test Plan:
[PASS] in 21.12 add an assorted network configuration with
       - vlan, bonded, and ethernet interfaces (the vlan interface on
         top of both base interfaces)
       - configure the bond interface with static address
       - add routes on all platform and data interfaces
[PASS] execute lock/unlock in 21.12 to verify config in applied
[PASS] execute AIO-SX upgrade to 22.12, at the bootstrap end check
        files in /etc/network/
[PASS] finish upgrade and after unlock verify the network access and
        address, interfaces, and route creation in the kernel

Closes-Bug: 2036451

Change-Id: I33e365a5900617e9aa9b7dabaaf9b7f467ac1e99
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>