Issue registry and RESTAPI/GUI certs in bootstrap
Include code to issue REST API & Web Server certificate
(system-restapi-gui-certificate) and Docker Registry certificate
(system-registry-local-certificate) using the 'system-local-ca'
ClusterIssuer in the bootstrap playbook.

Once the feature is completed and the control flag for it is removed,
this change will make HTTPS enabled out of the box in any StarlingX
deployment. Users will be advised to provide their RCA/ICAs to issue
these certs in new deployments, as well as to perform cert-manager
migration before upgrading to the release that will include this
feature (in future doc changes to the release notes).

- Included code to support upgrades. DX systems will require an upgrade
script that will be included in a parent review in 'config' repo. These
changes will be tested together with the ones for DC systems, to be
addressed in a future task in this Story. This upgrade code will be
dormant until that is concluded.

- Include a minor fix in certificates template to include 'ca.crt'
when provided through 'system_root_ca_cert' variable. This can be
tested by performing cert-manager migration.

The feature flag 'create_platform_certificates' will prevent
interfering with current logic while this feature isn't complete.

Test plan:
PASS: Deploy AIO-SX providing the CA cert for 'system-local-ca', as
      well as the certificate SANs data. Verify:
      - HTTPS is enabled and openstack public endpoints change into it
        after unlocking the controller.
      - The target certificates are issued by 'system-local-ca', and
        are managed by cert-manager.
      - The certificates in /etc/ssl/private are correct.
      - It's possible to log into the local Docker Registry.
      - Horizon is working as expected.

PASS: Perform cert-manager migration providing RCA and ICA. Check that
      the certificates/secrets are correct.

Story: 2009811
Task: 44802
Depends-on: https://review.opendev.org/c/starlingx/ansible-playbooks/+/896165

Change-Id: I281bbf42e7198ecb56ed7b3b247a091d36bf350d
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>