Disallow rehoming if there is a management affecting alarm
After introducing the kube rootca update as a task in the subcloud
rehoming, the subcloud rehoming should be fail in validation before
the actual process based on the requirement of the kube rootca update
strategy.

This commit adds the pre-check of the mgmt affecting alarms in the
subcloud before the acutal operation in the subcloud rehoming
playbook, removes the check of the online check of the controller-1
and the check of the 250.001 alarm because both of them can be
verified by the mgmt affecting alarm check.

Test plan:
1. Test rehoming SX/DX subcloud w/o mgmt affecting alarms successfully
- passed
2. Test rehoming DX subcloud with the controller-1 powered off,
rehoming failed as mgmt affecting alarm raised by the missing of the
standby services as expected, and the error message can be found in
the ansible log - passed.
3. Introducing mgmt affecting alarms [250.001 and 260.001] in the
subcloud, test rehoming failed as expected with the error message in
the ansible log - passed.

Closes-bug: 2038677
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I7a8fc37a189be1290a024b0127092d7ed33da484