Static routes aren't deleted after rehoming a subcloud
This commit fixes the issue of not deleting the previous
routes created on a subcloud after its being rehomed.

Test Plan:

PASS: Deploy a dc system with one controller and one
subcloud. Check the route to the system controller by
running "system host-route-list controller-0" in the
created subcloud.
Shutdown the previously created system controller and
create a new system controller with different mgmt
network configuration.
Rehome the subcloud to the new controller and check
that the route has changed to the new controller
route.

Closes-bug: 2039593

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: Ibee0af9463092348578565bc489aa2025b00b0d5