Reduce keyring access to once per playbook run.
The push-docker-images role uses a python script to
download, tag, and push images to the local registry.
There is a multi-threaded function which does this for
as many images as are needed.  For each job, the local
registry credentials are accessed.  Sometimes, for whatever
reason, the get_local_registry_auth function returns
None, a None type is passed to the docker client and
the routine fails.

This modification minimizes this to a single credential
access. Credentials are passed to each job as an input.

Test Plan:

PASS:
1. Install AIO-SX at lower STX version.
2. Initiate upgrade to higher STX version.
3. Before running system upgrade ansible playbook, update
   download_images.py with code in this review.
4. Run system upgrade playbook.
5. Observe no errors downloading or pushing images to local
   registry in sysinv.log, user.log or platform.log

Closes-Bug: 2039875
Change-Id: I1ec1dc88f72b38bdb0561ffb25f4e253441b094c
Signed-off-by: Joshua Reed <joshua.reed@windriver.com>