Increase retries for checking platform-integ-apps removal during restore
To ensure that the platform-integ-apps was removed over restore, there
is a step which waits for the platform-integ-apps to be removed
successfully before finishing it. This task is executed only when
wipe_ceph_osds=true.

The removal of platform-integ-apps is needed because FluxCD only
reapplies if its version has changed.

On slower machines, this step was failing because it was taking more
than 10 minutes to remove platform-integ-apps.

As it uncertain how much time will take to remove platform-integ-apps,
the number of retries was increased and now it takes at least 30 minutes
before failing. The idea is to have a longer timeout before failing, but
as soon as this step is successful it will exit this step and go to the
next one. The delay between retries was not changed because the command
execution takes about 7 seconds, which seems enough time to try again.

TEST PLAN:
PASS: Set wipe_ceph_osds=true and restore_mode=optimized in host
    overrides and verify that AIO-SX restore completes successfully with
    platform-integ-apps removed before the unlock and reapplied post
    unlock.
PASS: Set wipe_ceph_osds=false and restore_mode=optimized in host
    overrides and verify that AIO-SX restore completes successfully with
    platform-integ-apps in applied state before and after host unlock.

PASS: Set wipe_ceph_osds=true in host overrides and verify that AIO-SX
    restore completes successfully with platform-integ-apps removed
    before the unlock and reapplied post unlock.
PASS: Set wipe_ceph_osds=false in host overrides and verify that AIO-SX
    restore completes successfully with platform-integ-apps in applied
    state before and after host unlock.

PASS: Set wipe_ceph_osds=true in host overrides and verify that AIO-DX
    restore completes successfully with platform-integ-apps removed
    before the unlock and reapplied post unlock.
PASS: Set wipe_ceph_osds=false in host overrides and verify that AIO-DX
    restore completes successfully with platform-integ-apps in applied
    state before and after host unlock.

Partial-Bug: #2035115

Change-Id: I9078867736019da15892af46b1c4ad4e2893e861
Signed-off-by: Guilherme Costa <guilherme.costa@windriver.com>