Restart mtcAgent after upgrading CentOS networking
During stx6 to stx8 upgrade, the CentOS networking is upgraded using
a puppet runtime manifest.  During the application of this manifest,
there is the possibility of mtcAgent going down depending on the
network configuration. To unlock the node at the end of the upgrade
playbook mtcAgent is required.

To avoid being unable to unlock mtcAgent is restarted after the
networking runtime manifest is applied.

TEST PLAN
PASS: stx6 to stx8 AIO-SX upgrade on unaffected system
PASS: stx6 to stx8 AIO-SX upgrade on unaffected system

Closes-Bug: 2041194
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I49f60bcf452fb2e1b5e1a497097f5310b5e40b17