Remove k8s rootca update from rehome playbook
This commit removes the kube rootca update role from the rehome
playbook to decouple it from the rehome operation. This is done to
reduce risk as the update could fail due to many factors and recovering
a subcloud that failed rehoming is not straighforward.

The subcloud will be fully functional after the rehome process, even
without the k8s root CA update. However, the kube-rootca_sync_status
will remain 'in-sync' due to the current DC audit implementation.
Enhancements to DC audit, to ensure the transition to 'out-of-sync',
will be addressed in a separate commit.

Test Plan:
1. PASS - Execute subcloud rehome/migration and confirm the successful
          completion of the operation. Ensure that the kube root CA
          update role is not executed during this process.

Commit that added the k8s rootca update role (for reference):
https://review.opendev.org/c/starlingx/ansible-playbooks/+/876964

Story: 2010852
Task: 49082

Change-Id: I33fa189f565edb0b7ad738ae65923d2865373bd2
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>